07262021 Andrea Rudimentary NavBar
