+++ src/components/Home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Home = () => {
+  return (
+    <div>
+      
+    </div>
+  );
+}
+
+export default Home;
