+++ src/components/index.js
@@ -1,1 +1,2 @@
-export { default as App } from './App';
+export { default as Home } from './Home';
+export { default as NavbarComp } from './NavbarComp'
